Barcode input method works. Start cleaning.
