--- pos_barcode_tare/__init__.py
@@ -1,1 +1,0 @@
-from . import models
