--- pos_barcode_tare/i18n/fr_CH.po
@@ -1,185 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* pos_barcode_tare
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 9.0c\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-02-14 06:44+0000\n"
-"PO-Revision-Date: 2020-02-14 06:44+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"Language: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/xml/pos_barcode_tare.xml:8
-#, python-format
-msgid "Back"
-msgstr "Retour"
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/js/pos_barcode_tare.js:130
-#, python-format
-msgid "Failed to read weight from scale."
-msgstr "Echec lors de la lecture du poids depuis la balance"
-
-#. module: pos_barcode_tare
-#: model:product.uom,name:pos_barcode_tare.product_uom_kgm_dbg
-msgid "Kg"
-msgstr "Kg"
-
-#. module: pos_barcode_tare
-#: model:ir.ui.view,arch_db:pos_barcode_tare.view_pos_config_form
-msgid "Loose good options"
-msgstr "Options pour la vente en vrac"
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/js/pos_barcode_tare.js:196
-#, python-format
-msgid "Maximum tare weight is 99.999kg"
-msgstr "Le poids mesuré est supérieur au poids maximum pour la tare: 99.999kg"
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/xml/pos_barcode_tare.xml:18
-#, python-format
-msgid "Print"
-msgstr "Imprimer l'étiquette"
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/xml/pos_barcode_tare.xml:11
-#: code:addons/pos_barcode_tare/static/src/xml/pos_barcode_tare.xml:41
-#, python-format
-msgid "Print a tare label"
-msgstr "Imprimer une étiquette de tare"
-
-#. module: pos_barcode_tare
-#: model:ir.model.fields,help:pos_barcode_tare.field_pos_config_iface_tare_label
-msgid "Print tare labels with this POS"
-msgstr "Utiliser ce POS pour imprimer les étiquettes de tare"
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/xml/pos_barcode_tare.xml:15
-#, python-format
-msgid "Set the pot on the scale and check the weight above"
-msgstr ""
-"Placez le récipient à tarer sur la balance puis vérifiez le poids ci-dessus"
-
-#. module: pos_barcode_tare
-#: model:ir.model.fields,field_description:pos_barcode_tare.field_pos_config_iface_tare_label
-msgid "Show tare label button"
-msgstr "Utiliser ce POS pour imprimer les étiquettes de tare"
-
-#. module: pos_barcode_tare
-#: code:addons/pos_barcode_tare/models/barcode_rule.py:13
-#, python-format
-msgid "Tare"
-msgstr "Tare"
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/js/pos_barcode_tare.js:164
-#, python-format
-msgid ""
-"The scale sent a measure in %s unit. This unit of measure (UOM) in not found "
-"in the point of sale. You may need to create a new UOM named %s. The UOM "
-"name is case sensitive."
-msgstr ""
-"La balance à envoyé une mesure en %s. Cette unité de mesure (UOM) n'existe "
-"pas. Vous devez probablement créer une UOM nommée %s. Le nom de l'UOM est "
-"sensible à la casse."
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/js/pos_barcode_tare.js:288
-#, python-format
-msgid ""
-"The tare (%s) is already set for the product \"%s\". We can not re-apply a "
-"tare to this product."
-msgstr ""
-"La tare (%s) est déjà mesurée pour le produit \"%s\". Nous ne pouvons plus "
-"appliquer de tare à ce produit."
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/js/pos_barcode_tare.js:319
-#, python-format
-msgid ""
-"The tare weight is %s %s, it's greater or equal to the product weight %s. We "
-"can not apply this tare."
-msgstr ""
-"Le poids de la tare est de %s %s, c'est plus que le poids du produit %s. "
-"Nous ne pouvons pas appliquer cette tare."
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/js/pos_barcode_tare.js:74
-#, python-format
-msgid "We can not apply this tare barcode."
-msgstr "Impossible d'appliquer cette étiquette de tare."
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/js/pos_barcode_tare.js:33
-#, python-format
-msgid "We can not cast a weight in %s into %s."
-msgstr "Nous ne pouvons pas convertir un poids en %s en %s."
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/xml/pos_barcode_tare.xml:69
-#, python-format
-msgid "With a"
-msgstr "Avec un(e)"
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/xml/pos_barcode_tare.xml:61
-#, python-format
-msgid "at"
-msgstr "à"
-
-#. module: pos_barcode_tare
-#: model:ir.model,name:pos_barcode_tare.model_barcode_rule
-msgid "barcode.rule"
-msgstr "barcode.rule"
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/xml/pos_barcode_tare.xml:71
-#, python-format
-msgid "discount"
-msgstr "remise"
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/xml/pos_barcode_tare.xml:32
-#, python-format
-msgid "kg"
-msgstr "kg"
-
-#. module: pos_barcode_tare
-#: model:ir.model,name:pos_barcode_tare.model_pos_config
-msgid "pos.config"
-msgstr "pos.config"
-
-#. module: pos_barcode_tare
-#. openerp-web
-#: code:addons/pos_barcode_tare/static/src/xml/pos_barcode_tare.xml:32
-#: code:addons/pos_barcode_tare/static/src/xml/pos_barcode_tare.xml:78
-#, python-format
-msgid "tare ="
-msgstr "tare ="
-
-#~ msgid "Weight"
-#~ msgstr "Poids"
