--- pos_barcode_tare/models/pos_config.py
@@ -1,9 +1,0 @@
-from odoo import models, fields
-
-
-class PosConfig(models.Model):
-    _inherit = 'pos.config'
-    iface_tare_label = fields.Boolean(
-        'Show tare label button',
-        help="Print tare labels with this POS"
-    )
