--- pos_barcode_tare/pos_barcode_tare.xml
@@ -1,8 +1,0 @@
-<odoo>
-    <template id="assets_backend" name="pos_barcode_tare" inherit_id="point_of_sale.assets">
-            <xpath expr="." position="inside">
-                 <link rel="stylesheet" href="/pos_barcode_tare/static/src/css/pos_barcode_tare.css"/>
-                <script type="text/javascript" src="/pos_barcode_tare/static/src/js/pos_barcode_tare.js"></script>
-            </xpath>
-    </template>
-</odoo>
