--- pos_barcode_tare/readme/CONTRIBUTORS.rst
@@ -1,4 +1,0 @@
-- Le Nid
-- Sylvain LE GAL
-- CoopITEasy
-
