--- pos_barcode_tare/readme/ROADMAP.rst
@@ -1,1 +1,0 @@
-- Merge with Sylvain LE GAL tare module that allows user to set tare on the scale screen.
