--- pos_barcode_tare/views/pos_config_view.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0"?>
-
-
-<odoo>
-    <record model="ir.ui.view" id="view_pos_config_form">
-        <field name="name">view.pos.config.form</field>
-        <field name="model">pos.config</field>
-        <field name="inherit_id" ref="point_of_sale.pos_config_view_form" />
-        <field name="priority" eval="30"/>
-        <field name="arch" type="xml">
-            <div id="posbox_reference" position="inside">
-                <div class="col-12 col-lg-6 o_setting_box">
-                   <div class="o_setting_left_pane">
-                       <field name="iface_tare_label"/>
-                   </div>
-                   <div class="o_setting_right_pane">
-                      <label for="iface_tare_label"/>
-                         <div class="text-muted">
-                            Show tare barcode printer button on top of the num pad.
-                         </div>
-                      </div>
-                   </div>
-                </div>
-        </field>
-    </record>
-</odoo>
