+++ pos_tare/__init__.py
@@ -1,0 +1,1 @@
+from . import models
