+++ pos_tare/__manifest__.py
@@ -1,20 +1,24 @@
-# Copyright (C) 2015-Today GRAP (http://www.grap.coop)
-# @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
+# @author: François Kawala
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
 {
-    "name": "Point Of Sale - Tare",
-    "summary": "Manage Tare in Point Of Sale module",
-    "version": "12.0.1.0.0",
-    "category": "Point Of Sale",
-    "author": "GRAP, "
-              "Odoo Community Association (OCA)",
-    "maintainers": ["legalsylvain"],
-    "website": "https://www.github.com/OCA/pos",
-    "license": "AGPL-3",
-    "depends": ["point_of_sale"],
-    "data": ["views/assets.xml"],
-    "qweb": ["static/src/xml/pos_tare.xml"],
-    "installable": True,
-    "images": ["static/description/pos_tare.png"],
+    'name': "Point of Sale - Tare barecode labels for loose goods",
+    'version': '12.0.1.0.0',
+    'category': 'Point of Sale',
+    'summary': """Point of Sale - Print and scan tare \
+                  barecodes labels to sell loose goods.""",
+    'author': "Le Nid, Odoo Community Association (OCA)",
+    'website': "https://github.com/OCA/pos",
+    'license': 'AGPL-3',
+    'maintainers': ['fkawala'],
+    'depends': ['point_of_sale'],
+    'demo': ['demo/pos_tare_demo.xml'],
+    'data': [
+        'pos_tare.xml',
+        'views/pos_config_view.xml',
+        'data/barcode_rule.xml',
+    ],
+    'qweb': [
+        'static/src/xml/pos_tare.xml',
+    ],
+    'installable': True,
 }
