+++ pos_tare/data/barcode_rule.xml
