+++ pos_tare/demo/pos_tare_demo.xml
