--- pos_tare/i18n/fr.po
@@ -1,51 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* pos_tare
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 12.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-01-08 15:03+0000\n"
-"PO-Revision-Date: 2020-01-08 15:03+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: pos_tare
-#. openerp-web
-#: code:addons/pos_tare/static/src/xml/pos_tare.xml:11
-#, python-format
-msgid "Gross Weight"
-msgstr "Poids Brut"
-
-#. module: pos_tare
-#. openerp-web
-#: code:addons/pos_tare/static/src/js/screens.js:49
-#, python-format
-msgid "Incorrect Tare Value"
-msgstr "Valeur de tare incorrecte"
-
-#. module: pos_tare
-#. openerp-web
-#: code:addons/pos_tare/static/src/xml/pos_tare.xml:19
-#, python-format
-msgid "Kg"
-msgstr "Kg"
-
-#. module: pos_tare
-#. openerp-web
-#: code:addons/pos_tare/static/src/js/screens.js:50
-#, python-format
-msgid "Please set a numeric value in the tare field, or let empty."
-msgstr "Veuillez saisir une valeur numérique dans le champ 'tare', ou le laisser vide."
-
-#. module: pos_tare
-#. openerp-web
-#: code:addons/pos_tare/static/src/xml/pos_tare.xml:15
-#, python-format
-msgid "Tare"
-msgstr "Tare"
