+++ pos_tare/models/barcode_rule.py
