+++ pos_tare/pos_tare.xml
@@ -1,0 +1,8 @@
+<odoo>
+    <template id="assets_backend" name="pos_tare" inherit_id="point_of_sale.assets">
+            <xpath expr="." position="inside">
+                 <link rel="stylesheet" href="/pos_tare/static/src/css/pos_tare.css"/>
+                <script type="text/javascript" src="/pos_tare/static/src/js/pos_tare.js"></script>
+            </xpath>
+    </template>
+</odoo>
