--- pos_tare/readme/CONFIGURE.rst
@@ -1,2 +1,0 @@
-To enable this feature, you should enable "IoT Box" on your Point of Sale,
-and check the box "Electronic Scale"
