+++ pos_tare/readme/CONTRIBUTORS.rst
@@ -1,1 +1,4 @@
-* Sylvain LE GAL (https://www.twitter.com/legalsylvain)
+- Le Nid
+- Sylvain LE GAL
+- CoopITEasy
+
