+++ pos_tare/readme/ROADMAP.rst
@@ -1,2 +1,1 @@
-* Display the gross weight and the tare on each order line, on the screen
-  and on the receipt.
+- Merge with Sylvain LE GAL tare module that allows user to set tare on the scale screen.
