+++ pos_tare/readme/USAGE.rst
