+++ pos_tare/static/src/js/pos_tare.js
@@ -88,10 +88,16 @@ odoo.define('pos_barcode_tare.screens', function (require) {
             },
             // Setup the callback action for the "weight" barcodes.
             show: function () {
+                
+                console.log("okok");
                 this._super();
-                this.pos.barcode_reader.set_action_callback(
-                    'tare',
-                    _.bind(this.barcode_tare_action, this));
+                if (this.pos.config.iface_tare_method !== 'Manual') {
+                
+                    console.log("okokokok");
+                    this.pos.barcode_reader.set_action_callback(
+                        'tare',
+                        _.bind(this.barcode_tare_action, this));
+                    }
             },
         });
 
@@ -105,181 +111,92 @@ odoo.define('pos_barcode_tare.screens', function (require) {
         },
     });
 
-    screens.define_action_button({
-        'name': 'tareScreenButton',
-        'widget': TareScreenButton,
-    });
+    screens.ScaleScreenWidget.include({
 
-    // This is a new screen that reads weight from the electronic scale and
-    // create a barcode label encoding the weight. The screen shows a preview
-    // of the label. The user is expected to check if the preview matches what's
-    // measured on the scale. The barcode image is generated by the report
-    // module.
-    var TareScreenWidget = screens.ScreenWidget.extend({
-        template: 'TareScreenWidget',
-        next_screen: 'products',
-        previous_screen: 'products',
-        default_tare_value: 0.0,
-        weight_barcode_prefix: null,
+        // /////////////////////////////
+        // Overload Section
+        // /////////////////////////////
 
-        show: function () {
+        // Overload show function
+        // add an handler on the 
+        show: function(){
             this._super();
-            // Fetch the unit of measure used to save the tare
-            this.kg_unit = get_unit(this.pos, "kg");
-            // Fetch the barcode prefix from POS barcode parser rules.
-            this.weight_barcode_prefix = this.get_barcode_prefix();
-            // Setup the proxy
-            var queue = this.pos.proxy_queue;
-            // The pooling of the scale starts here.
+            this.tare = 0.0;
             var self = this;
-            queue.schedule(function () {
-                return self.pos.proxy.scale_read().then(function (weight) {
-                    try {
-                        self.set_weight(weight);
-                    } catch (error) {
-                        var title = _t("Failed to read weight from scale.");
-                        var popup = {title: title, body: error.message};
-                        self.gui.show_popup('error', popup);
-                    }
+            this.$('#input_weight_tare').keyup(function(event){
+                self.onchange_tare(event);
+            });
+            this.$('#input_weight_tare').focus()
+        },
+
+        // Overload set_weight function
+        // We assume that the argument is now the gross weight
+        // we compute the net weight, depending on the tare and the gross weight
+        // then we call super, with the net weight
+        set_weight: function(gross_weight){
+            this.gross_weight = gross_weight;
+            var net_weight = gross_weight - (this.tare || 0);
+            this.$('#container_weight_gross').text(this.get_product_gross_weight_string());
+            this._super(net_weight);
+        },
+
+        order_product: function(){
+            // TODO Set a warning, if the value is incorrect;
+            if (this.tare === undefined) {
+                this.gui.show_popup('error',{
+                    'title': _t('Incorrect Tare Value'),
+                    'body': _t('Please set a numeric value in the tare field, or let empty.'),
                 });
-            }, {duration:150, repeat: true});
-            // Shows a barcode whose weight might be zero, but this is preferred
-            // for UI/UX reasons.
-            this.render_receipt();
-            this.lock_screen(true);
-        },
-        get_barcode_prefix: function () {
-            var barcode_pattern = this.get_barcode_pattern();
-            return barcode_pattern.substr(0, 2);
-        },
-        get_barcode_pattern: function () {
-            var rules = this.get_barcode_rules();
-            var rule = rules.filter(
-                function (r) {
-                    // We select the first (smallest sequence ID) barcode rule
-                    // with the expected type.
-                    return r.type === tare_barcode_type;
-                })[0];
-            return rule.pattern;
-        },
-        get_barcode_rules: function () {
-            return this.pos.barcode_reader.barcode_parser.nomenclature.rules;
-        },
-        set_weight: function (scale_measure) {
-            var weight = scale_measure.weight;
-            var unit = get_unit(this.pos, scale_measure.unit);
-
-            if (typeof unit === 'undefined') {
-                throw new Error(_.str.sprintf(
-                    _t("The scale sent a measure in %s unit. This unit of "+
-                     "measure (UOM) in not found in the point of sale. You " +
-                     "may need to create a new UOM named %s. The UOM name is "+
-                     "case sensitive."), scale_measure.unit,
-                    scale_measure.unit));
             }
-
-            if (weight > 0) {
-                this.weight_in_kg = convert_mass(weight, unit, this.kg_unit);
-                this.render_receipt();
-                this.lock_screen(false);
-            }
-        },
-        get_weight: function () {
-            if (typeof this.weight_in_kg === 'undefined') {
-                return this.default_tare_value;
-            }
-            return this.weight_in_kg;
-        },
-        barcode_data: function (weight) {
-            // We use EAN13 barcode, it looks like 07 00000 12345 x. First there
-            // is the prefix, here 07, that is used to decide which type of
-            // barcode we're dealing with. A weight barcode has then two groups
-            // of five digits. The first group encodes the product id. Here the
-            // product id is 00000. The second group encodes the weight in
-            // grams. Here the weight is 12.345kg. The last digit of the barcode
-            // is a checksum, here symbolized by x.
-            var padding_size = 5;
-            var void_product_id = '0'.repeat(padding_size);
-            var weight_in_gram = weight * 1e3;
-
-            if (weight_in_gram >= Math.pow(10, padding_size)) {
-                throw new RangeError(_t("Maximum tare weight is 99.999kg"));
-            }
-
-            // Weight has to be padded with zeros.
-            var weight_with_padding = '0'.repeat(padding_size) + weight_in_gram;
-            var padded_weight = weight_with_padding.substr(
-                weight_with_padding.length - padding_size);
-            // Builds the barcode using a placeholder checksum.
-            var barcode = this.weight_barcode_prefix
-                .concat(void_product_id, padded_weight)
-                .concat(0);
-            // Compute checksum
-            var barcode_parser = this.pos.barcode_reader.barcode_parser;
-            var checksum = barcode_parser.ean_checksum(barcode);
-            // Replace checksum placeholder by the actual checksum.
-            return barcode.substr(0, 12).concat(checksum);
-        },
-        get_barcode_data: function () {
-            return this.barcode_data(this.get_weight());
-        },
-        lock_screen: function (locked) {
-            this._locked = locked;
-            if (locked) {
-                this.$('.print-label').addClass('disabled');
-            } else {
-                this.$('.print-label').removeClass('disabled');
-            }
-        },
-        print_web: function () {
-            window.print();
-            // TODO check this
-            this.pos.get_order()._printed = true;
-        },
-        print: function () {
-            // See comment in print function of ReceiptScreenWidget
-            this.lock_screen(true);
-            var self = this;
-            setTimeout(function () {
-                self.lock_screen(false);
-            }, 1000);
-
-            this.print_web();
-            this.click_back();
-        },
-        click_back: function () {
-            this.close();
-            this.gui.show_screen(this.previous_screen);
-        },
-        renderElement: function () {
-            this._super();
-            var self = this;
-            this.$('.back').click(function () {
-                self.click_back();
-            });
-            this.$('.print-label').click(function () {
-                if (!self._locked) {
-                    self.print();
+            else {
+                this._super();
+                if (this.tare > 0.0) {
+                    var order = this.pos.get_order();
+                    var orderline = order.get_last_orderline();
+                    orderline.set_tare(this.tare);
                 }
-            });
+            }
         },
-        render_receipt: function () {
-            this.$('.pos-tare-label-container').html(
-                QWeb.render('PosTareLabel', {widget:this}));
+
+        // /////////////////////////////
+        // Custom Section
+        // /////////////////////////////
+        get_product_gross_weight_string: function(){
+            var product = this.get_product();
+            var defaultstr = (this.gross_weight || 0).toFixed(3) + ' Kg';
+            if(!product || !this.pos){
+                return defaultstr;
+            }
+            var unit_id = product.uom_id;
+            if(!unit_id){
+                return defaultstr;
+            }
+            var unit = this.pos.units_by_id[unit_id[0]];
+            var weight = round_pr(this.gross_weight || 0, unit.rounding);
+            var weightstr = weight.toFixed(Math.ceil(Math.log(1.0/unit.rounding) / Math.log(10) ));
+            weightstr += ' ' + unit.name;
+            return weightstr;
         },
-        close: function () {
-            this._super();
-            delete this.weight;
-            this.pos.proxy_queue.clear();
+
+        onchange_tare: function(event){
+            this.tare = this.check_sanitize_value('#input_weight_tare');;
+            this.set_weight(this.gross_weight);
         },
-        get_tare_str: function () {
-            return format_tare(this.pos, this.get_weight(),
-                get_unit(this.pos, "kg"));
+
+        check_sanitize_value: function (input_name){
+            var res = this.$(input_name)[0].value.replace(',', '.').trim();
+            if (isNaN(res)){
+                this.$(input_name).css("background-color", "#F66");
+                return undefined;
+            }
+            else{
+                this.$(input_name).css("background-color", "#FFF");
+                return parseFloat(res, 10);
+            }
         },
+
     });
-
-    gui.define_screen({name:'tare', widget: TareScreenWidget});
-
+    
     // Update Orderline model
     var _super_ = models.Orderline.prototype;
     var OrderLineWithTare = models.Orderline.extend({
@@ -360,7 +277,6 @@ odoo.define('pos_barcode_tare.screens', function (require) {
 
     models.Orderline = OrderLineWithTare;
 
-    return {TareScreenWidget: TareScreenWidget,
-        OrderLineWithTare: OrderLineWithTare,
+    return {OrderLineWithTare: OrderLineWithTare,
         get_unit: get_unit};
 });
