+++ pos_tare/static/src/xml/pos_tare.xml
@@ -1,24 +1,60 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!--
-Copyright (C) 2015-Today GRAP (http://www.grap.coop)
-@author: Sylvain LE GAL (https://twitter.com/legalsylvain)
- License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
--->
 <templates id="template" xml:space="preserve">
+    
 
     <t t-extend="ScaleScreenWidget">
         <t t-jquery=".js-weight" t-operation="before">
-            <div class="weight-label">
-                Gross Weight
-            </div>
-            <div class="weight-value weight-value-fixed" id="container_weight_gross"/>
-            <div class="weight-label">
-                Tare
-            </div>
-            <div class="weight-value weight-value-alterable">
-                <input type="text" class="input-weight" id="input_weight_tare" placeholder="0.000"/> Kg
-            </div>
+            <t t-if="widget.pos.config.iface_tare_method != 'Barcode'">
+                <div class="weight-label">
+                    Gross Weight
+                </div>
+                <div class="weight-value weight-value-fixed" id="container_weight_gross"/>
+                <div class="weight-label">
+                    Tare
+                </div>
+                <div class="weight-value weight-value-alterable">
+                    <input type="text" class="input-weight" id="input_weight_tare" placeholder="0.000"/> Kg
+                </div>
+            </t>
         </t>
+    </t>
+    <t t-name="Orderline">
+        <li t-attf-class="orderline #{ line.selected ? 'selected' : '' }">
+            <span class="product-name">
+                <t t-esc="line.get_product().display_name"/>
+            </span>
+            <span class="price">
+                <t t-esc="widget.format_currency(line.get_display_price())"/>
+            </span>
+            <ul class="info-list">
+                <t t-if="line.get_quantity_str() !== '1' || line.selected ">
+                    <li class="info">
+                        <em>
+                            <t t-esc="line.get_quantity_str()" />
+                        </em>
+                        <t t-esc="line.get_unit().name" />
+                        at
+                        <t t-esc="widget.format_currency(line.get_unit_display_price(),'Product Price')" />
+                        /
+                        <t t-esc="line.get_unit().name" />
+                    </li>
+                </t>
+                <t t-if="line.get_discount_str() !== '0'">
+                    <li class="info">
+                        With a
+                        <em>
+                            <t t-esc="line.get_discount_str()" />%
+                        </em>
+                        discount
+                    </li>
+                </t>
+                <t t-if="line.get_tare_str() !== '0'">
+                    <li class="info">
+                        tare = <t t-esc="line.get_tare_str_with_unit()" />
+                    </li>
+                </t>
+            </ul>
+        </li>
     </t>
 
 </templates>
