--- pos_tare/views/assets.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-Copyright (C) 2015-Today GRAP (http://www.grap.coop)
-@author: Sylvain LE GAL (https://twitter.com/legalsylvain)
- License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
--->
-
-<odoo>
-
-    <template id="assets" inherit_id="point_of_sale.assets">
-        <xpath expr="." position="inside">
-            <script type="text/javascript" src="/pos_tare/static/src/js/screens.js"/>
-            <link rel="stylesheet" href="/pos_tare/static/src/css/pos_tare.css"/>
-        </xpath>
-    </template>
-
-</odoo>
